Move helper functions into spect/support and update rails_helper
